Rename schema files back to schema.ts Co-authored-by: jpaddison3 <jp@centreforeffectivealtruism.org>effectivealtruism.org>g>
