+++ packages/lesswrong/lib/collections/posts/collection.ts
@@ -1,4 +1,4 @@
-import schema from './temp_schema_2';
+import schema from './schema';
 import { createCollection } from '../../vulcan-lib';
 import { userOwns, userCanDo } from '../../vulcan-users/permissions';
 import { addUniversalFields, getDefaultResolvers } from '../../collectionUtils'
