+++ packages/lesswrong/lib/collections/posts/schema.ts
