+++ packages/lesswrong/lib/collections/sequences/schema.ts
